Use blue_velvet gem in place of Facebook module.
The Facebook connectivity portion of this app has always been written
with the intention of one day extracting its functionality into a
Rails engine. With blue_velvet, that dream has become a reality.

The blue_velvet gem extracts the functionality that is most unique
about thewonderbars.com and makes it available for any Rails developer
to use. It allows complete access to the developer for any Facebook
page it's assigned to, just like how the Wonder Bars' FacebookController
can access any portion of the page and even combine it into different
mixtures of information that are displayed on this web site in
creative ways.

As I am a proponent of "eat your own dog food", I have implemented the gem
in lieu of my existing native code into the app. All tests pass and everything
seems to be going smoothly.

blue_velvet: http://github.com/tubbo/blue_velvet

Changes:

- Build gemset with blue_velvet
- Rebuild CTags with new references
- Remove native Facebook::PageController
- Remove native Facebook::Client
- Remove duplicate Facebook views